CB-6212 iOS: fix warnings compiled under arm64 64-bit
(cherry picked from commit 007630da4866d867321008c6b8e1ebf57f05f259)

+++ src/ios/CDVCapture.m
@@ -423,7 +423,7 @@
         // NSLog(@"getFormatData: %@", [formatData description]);
     }
     if (bError) {
-        result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:errorCode];
+        result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:(int)errorCode];
     }
     if (result) {
         [self.commandDelegate sendPluginResult:result callbackId:callbackId];
